feat: first step to reorganize polkadot files
