+++ libs/coin-polkadot/src/bridge/deviceTransactionConfig.ts
