+++ libs/coin-polkadot/src/types/errors.ts
